Move `FeedImageDataLoaderCacheDecorator` to production
